Allow reuse of displays started by other users
Headless provides more than just the automatic launching of an Xvfb
process, it also provides useful screenshot and vidcap features. It
would be nice to be able to use these features even when reusing an Xvfb
display that's managed by another user.

With this change, reuse of another user's display will be allowed if a
display was explicitly given and reuse is explicitly allowed.

fixes #48